+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## a12
+
+- Mix-to-tileset code moved to TileSet.
+
+## Previous
+
+
+
 - MultiFixedConcentration allows a `min_volume` setting, which will raise an error of the minimum
   volume to be transferred for any component is too low.
 - Volume and concentration settings can now be changed, not just initialized, as strings.